PGSIM/requirements.txt
altgraph==0.17.3
aniso8601==9.0.1
asgiref==3.6.0
astroid==2.15.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
country-list==1.0.0
cycler==0.11.0
dill==0.3.6
distlib==0.3.6
Django==4.1.6
django-deep-serializer==0.1.3
exceptiongroup==1.1.1
filelock==3.9.0
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.39.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib==3.7.1
maven==0.0.6
mccabe==0.7.0
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pbr==5.11.1
pefile==2023.2.7
Pillow==9.5.0
pkinter==1.37.2
platformdirs==3.0.0
pluggy==1.0.0
protobuf==3.20.3
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pylint==2.17.1
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
pywin32-ctypes==0.2.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
stevedore==4.1.1
tomli==2.0.1
tomlkit==0.11.7
ttkthemes==3.2.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
utils==1.0.1
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.7
Werkzeug==2.2.2
windows-curses==2.3.1
wrapt==1.15.0
xlrd==2.0.1
